requirements_render.txt
# =============================================================================
# REQUIREMENTS POUR RENDER - Backend SAR
# =============================================================================
# Version optimisée pour le déploiement sur Render
# Équilibre entre fonctionnalités et performance

# =============================================================================
# CORE DJANGO - Dépendances essentielles
# =============================================================================
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
django-allauth==65.11.2

# =============================================================================
# BASE DE DONNÉES - PostgreSQL et extensions
# =============================================================================
psycopg2-binary==2.9.10
pgvector==0.4.1

# =============================================================================
# SERVEUR WEB - Production
# =============================================================================
gunicorn==21.2.0
whitenoise==6.6.0

# =============================================================================
# CONFIGURATION - Variables d'environnement
# =============================================================================
python-decouple==3.8
python-dotenv==1.1.1

# =============================================================================
# IMAGES ET MÉDIA - Traitement des fichiers
# =============================================================================
pillow==11.3.0

# =============================================================================
# AUTHENTIFICATION - OAuth Google
# =============================================================================
google-auth==2.41.0
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
oauthlib==3.3.1
requests-oauthlib==2.0.0

# =============================================================================
# RAG - Système d'assistance intelligent (optimisé)
# =============================================================================
# Modèle de vectorisation léger
sentence-transformers==2.7.0
# Recherche vectorielle optimisée
faiss-cpu==1.7.4
# Framework RAG
langchain==0.1.0
langchain-community==0.0.10
# Calculs numériques
numpy==1.24.3
# Hub Hugging Face
huggingface-hub==0.19.4

# =============================================================================
# UTILITAIRES - Fonctionnalités de base
# =============================================================================
# Parsing SQL
sqlparse==0.5.3
# Extensions de type
typing_extensions==4.15.0
# Données de fuseau horaire
tzdata==2025.2
# Validation de données
pydantic==2.11.9
pydantic-settings==2.10.1
# Requêtes HTTP
requests==2.32.5
httpx==0.28.1
# JSON optimisé
orjson==3.11.3
# Cache
cachetools==6.2.0

# =============================================================================
# SÉCURITÉ - Authentification et sécurité
# =============================================================================
PyJWT==2.10.1
cryptography==41.0.7

# =============================================================================
# DÉPENDANCES SYSTÈME - Support système
# =============================================================================
# Support asynchrone
asgiref==3.9.1
# Support des attrs
attrs==25.3.0
# Support des dataclasses
dataclasses-json==0.6.7
# Support des fichiers
filelock==3.19.1
# Support des URLs
yarl==1.20.1
# Support des signaux
aiosignal==1.4.0
# Support des types annotés
annotated-types==0.7.0
# Support des tâches asynchrones
anyio==4.10.0
# Support des certificats
certifi==2025.8.3
# Support de l'encodage
charset-normalizer==3.4.3
# Support des couleurs
colorama==0.4.6
# Support des listes gelées
frozenlist==1.7.0
# Support des systèmes de fichiers
fsspec==2025.9.0
# Support des tâches
greenlet==3.2.4
# Support HTTP
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx-sse==0.4.1
# Support des identifiants
idna==3.10
# Support des templates
Jinja2==3.1.6
# Support des tâches parallèles
joblib==1.5.2
# Support des patches JSON
jsonpatch==1.33
jsonpointer==3.0.0
# Support des langues
langsmith==0.4.29
# Support du markup
MarkupSafe==3.0.2
# Support des schémas
marshmallow==3.26.1
# Support des calculs
mpmath==1.3.0
# Support des dictionnaires multiples
multidict==6.6.4
# Support des extensions mypy
mypy_extensions==1.1.0
# Support des réseaux
networkx==3.5
# Support des packages
packaging==25.0
# Support des propriétés
propcache==0.3.2
# Support des modules ASN.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
# Support du parsing
pyparsing==3.2.5
# Support des expressions régulières
regex==2025.9.18
# Support des requêtes
requests-toolbelt==1.0.0
# Support RSA
rsa==4.9.1
# Support des tenseurs sécurisés
safetensors==0.6.2
# Support de l'apprentissage automatique
scikit-learn==1.7.2
scipy==1.16.2
# Support des ensembles
setuptools==80.9.0
# Support des sniffers
sniffio==1.3.1
# Support SQLAlchemy
SQLAlchemy==2.0.43
# Support des calculs symboliques
sympy==1.14.0
# Support de la persistance
tenacity==9.1.2
# Support des pools de threads
threadpoolctl==3.6.0
# Support des tokenizers
tokenizers==0.22.1
# Support des barres de progression
tqdm==4.67.1
# Support des transformers
transformers==4.56.2
# Support de l'inspection de types
typing-inspect==0.9.0
typing-inspection==0.4.1
# Support des URLs
urllib3==2.5.0
# Support de la compression
zstandard==0.25.0

# =============================================================================
# NOTES D'OPTIMISATION
# =============================================================================
# - torch n'est pas inclus (trop lourd pour Render)
# - sentence-transformers version 2.7.0 (plus légère)
# - faiss-cpu au lieu de faiss-gpu
# - numpy version stable 1.24.3
# - langchain version 0.1.0 (compatible)
# - Toutes les dépendances sont en versions stables
# - Optimisé pour un usage CPU uniquement
